requirements.txt
bcrypt==4.1.2
bidict==0.23.1
blinker==1.7.0
certifi==2025.4.26
click==8.1.7
cloudinary==1.44.0
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.0
flask==3.0.2
flask-cors==4.0.0
flask-marshmallow==1.2.0
flask-socketio==5.5.1
flask-sqlalchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
itsdangerous==2.1.2
jinja2==3.1.3
markupsafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
packaging==23.2
pymysql==1.1.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
simple-websocket==1.1.0
six==1.17.0
sqlalchemy==2.0.25
typing-extensions==4.9.0
urllib3==2.4.0
waitress==3.0.0
werkzeug==3.0.1
wsproto==1.2.0
